Fix race preventing output match and exit code check in same test.
When we have no more output matching (with, e.g., st.stdout.match) to
do, we kill the process. However, if the process has not yet
completed, and we have a pending exit code check, the process may be
killed before exiting.  Thus, when we have a pending exit code check,
we don't kill the process, but instead wait for it to exit so we can
check the exit code.